email: rework for package, change type definitions, add File
This patch expands on the comments for the Email type and subtypes. Some
type names have been changed to allow for use as an external package.

The separate inline and attached file types have been incorporated into
a single "File" type. The new file type exposes a number of new fields
including the underlying io.Reader.